Added the concept of source AARs (srcaar).
Unity will happily attempt to process all AARs included in a project
which is unfortunate since some AARs need to be expanded and have
variable replacement applied.  This change allows AARs to be included
in a project using the .srcaar extension which the Jar Resolver will
treat as typical AAR files, copying / expanding them as required.

Bug: 30595610
Change-Id: Ib74c2987bf0129a3beaac6de02819ebce4c0cbc1