Fixed incorrect update of unpacked AARs.
When AARs are unpacked into directories the .aar file extension
is stripped.  The extension was incorrectly being stripped from
unpacked directories when checking for updated versions of AARs
resulting in the last component of a version number being removed
e.g foo-1.2.3.aar would be unpacked as foo-1.2.3 and end up
being parsed as package foo-1.2 such that when checking for
updates of foo, version 1.2.3 would look like a new version vs.
1.2.

Bug: 30597815
Change-Id: I47bbb4c1ab3d6f556e8c1e4930c9eefc4dfd0590