Fixed upgrade to new version of packages.
If a package was already installed (i.e its maven repo was present)
it wouldn't be upgraded when a dependency required a newer version
of the package.  This fixes the issue by checking for a found
artifact for each dependency and then forcing a reinstall / upgrade
of the required package if the artifact isn't found.

Bug: 30760595
Change-Id: Ib2eb423b3da683b1c3ebc899672170eba10211ba