Fixed Play Services Resolver being disabled on import at the same version.
When Unity imports a package with the same GUIDs it will apply the metadata
from the newly imported package into the project.  In the case of the
Play Services Resolver, by default IOSResolver and JarResolver are disabled
forcing the user to manually re-enable the DLLs as the VersionResolver
will not see metadata usually encoded in the filename
(plugin_v1.2.3_tEditor.dll will have been renamed plugin.dll) instructing
it to enable the plugins.

This patch moves the target platform to labels of each component such that
an import will force the VersionHandler to re-enable the components when it
executes.

In addition, this adds the version number of each component into the file
metadata so that it isn't lost when imported over an existing plugin at the
same revision.

BUG=31812730

Change-Id: I92b6ac4269ad6f6454999adf15e77b3acbf68305