Fixed latest version selection when version specs are too strict.
If version specifications result in non-overlapping versions for a
dependency, no dependency will be found in which case the latest
version of the dependency should be selected.

This patch fixes latest version selection and adding logging to
describe which set of dependencies led to this fallback state.

Also, sprinkled some more verbose logging around to aid with
future debugging.

BUG=31845309

Change-Id: I5f319777285f4cfe386ea2514af404a58b118619