Attempt at fixing github issue #48
I was unable to reproduce
https://github.com/googlesamples/unity-jar-resolver/issues/48
however from the stack trace it looks like the cached target dependency
isn't found by GooglePlayServices.ResolverVer1_1.ShouldReplaceDependency().
This is very likely to be caused by the AAR cache referencing a file that
has since been deleted.

In the case the target file isn't found we simply permit the replacement
of the dependency.

Change-Id: I595e395a750362b38ed7a1a6f638a5751b45a87a