Linux library renaming based upon Unity version.
VersionHandler will now rename Linux libraries so they can target Unity
versions that require different file naming.  Libraries need to be labelled
gvh_linuxlibname-${basename} in order to be considered for renaming.
e.g gvh_linuxlibname-MyLib will be named MyLib.so in Unity 5.5 and below and
libMyLib.so in Unity 5.6 and above.

Also, changed VersionHandlerImpl so that it is only disabled via the command
line option -gvh_disable similar to VersionHandler (bootstrap) this should
reduce confusion when it doesn't run by default in batch mode.

Bug: 64066718
Change-Id: I16d80bfd532c0dc861327770a11adddd80380a3f